+++ app/src/main/kotlin/me/amrbashir/hijriwidget/preferences/MainActivity.kt
@@ -41,7 +41,6 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shadow
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.text.style.TextDirection
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.graphics.ColorUtils
@@ -56,6 +55,7 @@ import me.amrbashir.hijriwidget.Preferences
 import me.amrbashir.hijriwidget.PreferencesTheme
 import me.amrbashir.hijriwidget.android.AlarmReceiver
 import me.amrbashir.hijriwidget.isDark
+import me.amrbashir.hijriwidget.preferences.routes.CalendarCalculation
 import me.amrbashir.hijriwidget.preferences.routes.DayOffset
 import me.amrbashir.hijriwidget.preferences.routes.Home
 import me.amrbashir.hijriwidget.preferences.routes.Language
@@ -66,6 +66,7 @@ import me.amrbashir.hijriwidget.widget.HijriWidget
 object Route {
     const val HOME = "/"
     const val DAY_OFFSET = "DayOffset"
+    const val CALENDAR_CALCULATION_METHOD = "CalendarCalculationMethod"
     const val LANGUAGE = "Language"
     const val TEXT_COLOR = "TextColor"
     const val TEXT_SIZE = "TextSize"
@@ -193,6 +194,7 @@ open class WidgetConfiguration(private val autoClose: Boolean = true) : Componen
                         ) {
                             composable(Route.HOME) { Home() }
                             composable(Route.DAY_OFFSET) { DayOffset() }
+                            composable(Route.CALENDAR_CALCULATION_METHOD) { CalendarCalculation() }
                             composable(Route.LANGUAGE) { Language() }
                             composable(Route.TEXT_COLOR) { TextColor() }
                             composable(Route.TEXT_SIZE) { TextSize() }
@@ -230,6 +232,7 @@ open class WidgetConfiguration(private val autoClose: Boolean = true) : Componen
             Preferences.language.value,
             Preferences.dayStart.value,
             Preferences.dayOffset.value,
+            Preferences.calendarCalculationMethod.value,
             HijriDate.today.value
         ) {
             date = HijriDate.today()