+++ app/src/main/kotlin/me/amrbashir/hijriwidget/preferences/routes/CalendarCalculation.kt
@@ -6,26 +6,29 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import me.amrbashir.hijriwidget.CalendarCalculationMethod
 import me.amrbashir.hijriwidget.Preferences
-import me.amrbashir.hijriwidget.SupportedLanguage
 import me.amrbashir.hijriwidget.preferences.composables.PreferenceCategory
 import me.amrbashir.hijriwidget.preferences.composables.RadioIcon
 
 @Composable
-fun Language() {
-    val savedLang = Preferences.language.value
+fun CalendarCalculation() {
+    val savedCalcMethod = Preferences.calendarCalculationMethod.value
 
     Column(
         Modifier
             .fillMaxSize()
             .verticalScroll(rememberScrollState())
     ) {
-        for (lang in SupportedLanguage.entries) {
+        for (calcMethod in CalendarCalculationMethod.entries) {
+            val methodStr = calcMethod.toString()
+
             PreferenceCategory(
-                label = "$lang",
-                alternateIcon = { RadioIcon(selected = savedLang == lang) },
+                label = calcMethod.label(),
+                description = calcMethod.desc(),
+                alternateIcon = { RadioIcon(selected = savedCalcMethod == methodStr) },
                 onClick = {
-                    Preferences.language.value = lang
+                    Preferences.calendarCalculationMethod.value = methodStr
                 }
             )
         }